Add files via upload
As a member of the Reuhub team, I played a crucial role in optimizing their data pipelines to enhance their overall efficiency. One particular challenge that we faced was the slow runtime of certain Python transformation functions that were taking too long to execute. To overcome this issue, I suggested converting these functions into SQL transformations that could be executed inside the database itself.

To implement this solution, I worked closely with the team to identify the specific functions that were causing the delays and then re-wrote them in SQL. By doing so, we were able to minimize the data transfer between the database and Python, resulting in significantly faster execution times.

As a result of these efforts, we were able to decrease the runtime of the program by a staggering 75 percent. This improvement was crucial for the Reuhub team as it allowed them to process their data much faster, enabling them to make more informed decisions in a shorter amount of time.

Overall, my experience working with Reuhub was a rewarding one, and I am proud to have contributed to their success by optimizing their data pipelines through the use of Python, Airflow, and SQL.